Enable use of quorra as a vCenter extension
This is required to perform various operations that we want to do,
such as managing IP Pools.

Being an extension requires the use of SSL certificates for
authentication so there's now also some commands to generate them.

(Getting `LoginByExtensionCertificate` working was incredibly
painful as it's not really documented that you need to use a
proxied connection through vSphere to a private, internal, address.
Note: should really pull this out into a patch against govmomi.)